requirements.txt
Flask
openpyxl
python-pptx
Flask-Mail